+++ src/solver/flags.rs
@@ -1,5 +1,5 @@
 // src/lbm/check.rs
 // LBM FLAGS
-pub const FLAG_FLUID: i32 = 0;
-pub const FLAG_SOLID: i32 = 1;
-pub const FLAG_EQ: i32 = 2;
+pub const FLAG_FLUID: u8 = 0;
+pub const FLAG_SOLID: u8 = 1;
+pub const FLAG_EQ: u8 = 2;
